+++ packages/udaru-core/config.js
@@ -63,6 +63,11 @@ const defaultConfig = {
                 Effect: 'Allow',
                 Action: [Action.ReadPolicyVariables],
                 Resource: [resources.policies({ organizationId: ':organizationId' })]
+              },
+              {
+                Effect: 'Allow',
+                Action: [Action.ListPolicies],
+                Resource: [resources.policies({ organizationId: ':organizationId' })]
               }
             ]
           }
@@ -105,6 +110,7 @@ const defaultConfig = {
                   Action.UpdateTeam,
                   Action.AddTeamMember,
                   Action.ReplaceTeamMember,
+                  Action.ListTeamPolicies,
                   Action.RemoveTeamMember
                 ],
                 Resource: [