+++ packages/udaru-hapi-plugin/test/endToEnd/users.test.js
@@ -597,6 +597,75 @@ lab.experiment('Users - manage policies', () => {
     await udaru.users.replacePolicies({ id: result.id, policies: ['policyId2'], organizationId: result.organizationId })
   })
 
+  lab.test('List user policies', async () => {
+    let options = utils.requestOptions({
+      method: 'POST',
+      url: '/authorization/users/VerucaId/policies',
+      payload: {policies: [{
+        id: 'policyId2',
+        variables: {var1: 'value1'}
+      }, {
+        id: 'policyId2',
+        variables: {var2: 'value2'}
+      }, {
+        id: 'policyId2',
+        variables: {var3: 'value3'}
+      }]}
+    })
+
+    let response = await server.inject(options)
+    let result = response.result
+
+    expect(response.statusCode).to.equal(200)
+    expect(result.policies.length).to.equal(3)
+
+    options = utils.requestOptions({
+      method: 'GET',
+      url: '/authorization/users/VerucaId/policies'
+    })
+
+    response = await server.inject(options)
+    result = response.result
+
+    expect(response.statusCode).to.equal(200)
+    expect(result.total).to.equal(3)
+    expect(utils.PoliciesWithoutInstance(result.data)).to.contain([{
+      id: 'policyId2',
+      name: 'Accountant',
+      version: '0.1',
+      variables: {var1: 'value1'}
+    }])
+
+    options = utils.requestOptions({
+      method: 'GET',
+      url: '/authorization/users/VerucaId/policies?limit=100&page=1'
+    })
+
+    response = await server.inject(options)
+    result = response.result
+
+    expect(response.statusCode).to.equal(200)
+    expect(result.total).to.equal(3)
+    expect(utils.PoliciesWithoutInstance(result.data)).to.contain([ {
+      id: 'policyId2',
+      name: 'Accountant',
+      version: '0.1',
+      variables: {var2: 'value2'}
+    }])
+
+    await udaru.users.replacePolicies({ id: 'VerucaId', policies: ['policyId2'], organizationId: 'WONKA' })
+  })
+
+  lab.test('get non existent users policies', async () => {
+    const options = utils.requestOptions({
+      method: 'GET',
+      url: '/authorization/users/doesnotexist/policies?limit=100&page=1'
+    })
+
+    let response = await server.inject(options)
+    expect(response.statusCode).to.equal(404)
+  })
+
   lab.test('add policy with invalid ID to a user', async () => {
     const options = utils.requestOptions({
       method: 'PUT',